+++ src/components/settings/settings.module.css
@@ -24,3 +24,15 @@
 .settingsLabel {
 	margin-left: 4px;
 }
+
+.settingsButton {
+	display: block;
+	float: right;
+	margin: 12px;
+	padding: 8px 14px;
+	border: solid 1px rgb(16, 245, 104);
+	border-radius: 4px;
+	cursor: pointer;
+	background-color: rgb(16, 245, 104);
+	color: #fff;
+}
