+++ src/types/settings.ts
@@ -4,10 +4,6 @@ export type SettingsProps = {
 	settingsStore: SettingsStore;
 }
 
-export type SettingsRepositoryProps = {
-	items: string[];
-} & SettingsProps;
-
 export type SettingsSectionProps = {
 	heading: string;
 	items: string[];
